requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
boto3==1.14.14
botocore==1.17.14
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorclass==2.2.0
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.39.1
django-cors-headers==3.3.0
django-countries==6.1.2
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docopt==0.6.2
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
pep8==1.7.1
Pillow==7.1.2
pip-upgrader==1.4.15
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
terminaltables==3.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
